Retirada das conversoes de tipo

+++ server/api/controllers/resultset.js
@@ -553,13 +553,13 @@ function associaAgregacao(indicador){
     for(key in indicador.indicadores){
       //ans = ans.replace(`[${key}]`, `${operacao}(${key})::float`);
       // Para substituir todas as ocorrencias em javascript:
-      ans = ans.replace(new RegExp(`[${key}]`.replace(/([.*+?^=!:${}()|\[\]\/\\])/g, "\\$1"), 'g'), `${operacao}(${key})::float`);
+      ans = ans.replace(new RegExp(`[${key}]`.replace(/([.*+?^=!:${}()|\[\]\/\\])/g, "\\$1"), 'g'), `${operacao}(${key})`);
     }
     ans += ` ${indicador.codigo},`;
     //console.log('PARTICULA==>', ans);
   }else{
     // select += ` SUM(${key})::float ${key},`;
-    ans = ` ${operacao}(${indicador.codigo})::float ${indicador.codigo},`;
+    ans = ` ${operacao}(${indicador.codigo}) ${indicador.codigo},`;
   }
   return ans;
 }
